\encoding{utf8}
\name{arules}
\alias{arules}
\title{Association rules classification using the arules package.}
\description{
  Association rules classification using the arules
  package.
}
\references{
  Michael Hahsler, Christian Buchta, Bettina Grün and Kurt
  Hornik. Introduction to arules – A computational
  environment for mining association rules and frequent
  item sets
}

