\name{build.intervals}
\alias{build.intervals}
\title{Build the intervals given the cut points.}
\usage{
  build.intervals(cut.points)
}
\arguments{
  \item{cut.points}{vector with the points where split the
  values range.}
}
\value{
  List with an entry for each interval. Every interval is
  represented by a vector with lower and upper bound.
}
\description{
  Build the intervals given the cut points.
}

