\name{cacc.cut.points}
\alias{cacc.cut.points}
\title{Compute cut points.}
\usage{
  cacc.cut.points(values, classes)
}
\arguments{
  \item{values}{vector with the continuous values to
  discretize.}

  \item{classes}{vector with the classes associated with
  the values in \code{values}.}
}
\value{
  Vector with the cut points.
}
\description{
  Compute the cut points for the given values.
}

