\encoding{utf8}
\name{cacc}
\alias{cacc}
\title{Discretization algorithm based on Class-Attribute Contingency Coefficient}
\description{
  Discretization algorithm based on Class-Attribute
  Contingency Coefficient
}
\references{
  C.J. Tsai, C.-I. Lee, W.-P. Yang. A discretization
  algorithm based on Class-Attribute Contingency
  Coefficient. Information Sciences 178:3 (2008) 714-731.

  J. Alcalá-Fdez, L. Sánchez, S. García, M.J. del Jesus, S.
  Ventura, J.M. Garrell, J. Otero, C. Romero, J. Bacardit,
  V.M. Rivas, J.C. Fernández, F. Herrera. KEEL: A Software
  Tool to Assess Evolutionary Algorithms to Data Mining
  Problems. Soft Computing 13:3 (2009) 307-318, doi:
  10.1007/s00500-008-0323-y.

  J. Alcalá-Fdez, A. Fernandez, J. Luengo, J. Derrac, S.
  García, L. Sánchez, F. Herrera. KEEL Data-Mining Software
  Tool: Data Set Repository, Integration of Algorithms and
  Experimental Analysis Framework. Journal of
  Multiple-Valued Logic and Soft Computing 17:2-3 (2011)
  255-287.
}

