\name{cacc.value}
\alias{cacc.value}
\title{Compute the cacc value}
\usage{
  cacc.value(values, classes, cut.points)
}
\arguments{
  \item{values}{vector with the continuous values to
  discretize in ascending order.}

  \item{classes}{vector with the classes associated with
  the values in \code{values}.}

  \item{cut.points}{vector with the cut points.}
}
\value{
  CACC value for the given discretization scheme.
}
\description{
  Compute the cacc value
}

