\name{discretize.values}
\alias{discretize.values}
\title{Discretize a series of values using the given intervals}
\usage{
  discretize.values(values, intervals)
}
\arguments{
  \item{values}{vector with the continuous values to
  discretize.}

  \item{intervals}{list with an entry for each interval.
  Every interval must be represented by a vector with lower
  and upper bound.}
}
\value{
  Vector with the given values discretized.
}
\description{
  Discretize a series of values using the given intervals
}

