\name{ewib.cut.points}
\alias{ewib.cut.points}
\title{Compute cut points.}
\usage{
  ewib.cut.points(values, num.bin = 10)
}
\arguments{
  \item{values}{vector with the continuous values to
  discretize.}

  \item{num.bin}{number of bins}
}
\value{
  Vector with the cut points.
}
\description{
  Compute the cut points for the given values.
}

