\name{gdiscretize.id3}
\alias{gdiscretize.id3}
\title{Produce set of intervals.}
\usage{
  gdiscretize.id3(dataset)
}
\arguments{
  \item{dataset}{training set to be discretized}
}
\value{
  discretized dataset
}
\description{
  Produce set of intervals.
}
\references{
  J.R. Quinlan. Induction of Decision Trees. Machine
  Learning 1 (1986) 81-106.
}

