\name{id3.entropy.partition}
\alias{id3.entropy.partition}
\title{Entropy of a partition.}
\usage{
  id3.entropy.partition(classes, split.point)
}
\arguments{
  \item{classes}{vector with the attribute classes}

  \item{split.point}{index of the element that separates
  the partitions}
}
\value{
  Entropy of the partition.
}
\description{
  Entropy of a partition.
}

