\encoding{utf8}
\name{id3}
\alias{id3}
\alias{id3.discretize}
\title{ID3 Discretizer}
\usage{
  id3.discretize(values, classes)
}
\arguments{
  \item{values}{vector with the continuous values to
  discretize}

  \item{classes}{vector with the classes associated with
  the values in \code{values}}
}
\value{
  Vector with the \code{values} discretized
}
\description{
  ID3 Discretizer

  Discretize the values of an attribute.
}
\references{
  J.R. Quinlan. Induction of Decision Trees. Machine
  Learning 1 (1986) 81-106.

  J. Alcalá-Fdez, L. Sánchez, S. García, M.J. del Jesus, S.
  Ventura, J.M. Garrell, J. Otero, C. Romero, J. Bacardit,
  V.M. Rivas, J.C. Fernández, F. Herrera. KEEL: A Software
  Tool to Assess Evolutionary Algorithms to Data Mining
  Problems. Soft Computing 13:3 (2009) 307-318, doi:
  10.1007/s00500-008-0323-y.

  J. Alcalá-Fdez, A. Fernandez, J. Luengo, J. Derrac, S.
  García, L. Sánchez, F. Herrera. KEEL Data-Mining Software
  Tool: Data Set Repository, Integration of Algorithms and
  Experimental Analysis Framework. Journal of
  Multiple-Valued Logic and Soft Computing 17:2-3 (2011)
  255-287.
}

