\name{read}
\alias{read}
\title{Read a Tab Separated Values file}
\usage{
  read(filename)
}
\arguments{
  \item{filename}{filename from which read}
}
\value{
  dataframe
}
\description{
  Read a Tab Separated Values file
}
\details{
  Describe the format of the file
}
\references{
  \url{http://en.wikipedia.org/wiki/Tab-separated_values}
}

